Add Windows impl for async mmap prefetch
