+++ src/common/util/src/os/win/win_memory.cpp
@@ -9,19 +9,15 @@
 #include <malloc.h>
 #include <windows.h>
 
-#include <algorithm>
 #include <cassert>
 #include <cstddef>
 #include <cstring>
-#include <thread>
 #include <tuple>
 
 #include "memory_prefetch.hpp"
 #include "openvino/util/memory.hpp"
 
 namespace ov::util {
-
-void populate_pages(void* ptr, size_t size, size_t num_threads) noexcept;
 
 void* aligned_alloc(size_t size, size_t alignment) noexcept {
     if (alignment == 0) {
@@ -68,13 +64,18 @@ void vm_prefetch(void* ptr, size_t size, size_t num_threads) noexcept {
         WIN32_MEMORY_RANGE_ENTRY entry{ptr, size};
         ::PrefetchVirtualMemory(::GetCurrentProcess(), 1, &entry, 0);
     } else {
-        // blocks until every page has been faulted in.
-        populate_pages(ptr, size, num_threads);
+        PrefetchToken(submit_page_toucher_tasks(ptr, size, num_threads)).wait();
     }
 }
 
-PrefetchToken vm_prefetch_async(void* /*ptr*/, size_t /*size*/, size_t /*num_threads*/) noexcept {
-    return {};
+PrefetchToken vm_prefetch_async(void* ptr, size_t size, size_t num_threads) noexcept {
+    assert(ptr != nullptr && size > 0);
+    if (num_threads == 0) {
+        WIN32_MEMORY_RANGE_ENTRY entry{ptr, size};
+        ::PrefetchVirtualMemory(::GetCurrentProcess(), 1, &entry, 0);
+        return {};
+    }
+    return PrefetchToken(submit_page_toucher_tasks(ptr, size, num_threads));
 }
 
 }  // namespace ov::util
